pyproject.toml
[project]
name = "grafana-dashboards-generator"
version = "0.1.0"
description = "Dump and Generate Grafana Dashboards with jinja2 and yaml"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
]
